Add module-level doc comments
- Move timing functions to setup.rs
