+++ src/qed/mod.rs
@@ -1,3 +1,5 @@
+//! Rates and spectra for strong-field QED processes
+
 pub mod photon_emission;
 pub mod photon_absorption;
 mod pwmci;
